Refactor util functions
- Move `get_iap_garage` function to utils
- Separate common utils to multiple files
